homework 5 attempt 1
